Minor changes to toolchain
- Started work on building a custom toolchain for the OS; this is now
  used to build the OS, root server and the C library.

  We need to somehow actually start building up the sysroot (so we can
  find the C library and such when building other userspace programs)
  and need to figure out dynamic linking/ABI stuff.

+++ user/libc/src/string/memset.c
@@ -4,7 +4,7 @@
 /*
  * Fills a given segment of memory with a specified value.
  */
-void *memset(void *ptr, const uint8_t value, const size_t num) {
+void *memset(void *ptr, const int value, const size_t num) {
     if(value == 0x00) {
         return memclr(ptr, num);
     }
